backend/requirements.txt
Flask_SQLAlchemy==2.4.0
Flask_Migrate==2.4.0
alembic==1.0.10
SQLAlchemy==1.3.3
Flask_Script==2.0.6
Werkzeug==0.15.2
Flask_Login==0.4.1
Flask==1.0.2
WTForms==2.2.1
Flask_WTF==0.14.2
pymysql==0.9.3
python-dotenv==0.10.1
flask_uploads==0.2.1
flask_rq2==18.3
flask_paginate==0.5.2
